removed unused flow code
