+++ src/main/java/picard/flow/FlowBasedArgumentCollection.java
@@ -13,15 +13,12 @@ public class FlowBasedArgumentCollection implements Serializable {
     public static final String PROBABILITY_RATIO_THRESHOLD_LONG_NAME = "flow-probability-threshold";
     public static final String REMOVE_LONGER_THAN_ONE_INDELS_LONG_NAME = "flow-remove-non-single-base-pair-indels";
     public static final String REMOVE_ONE_TO_ZERO_PROBS_LONG_NAME = "flow-remove-one-zero-probs";
-    public static final String NUMBER_OF_POSSIBLE_PROBS_LONG_NAME = "flow-quantization-bins";
     public static final String FILLING_VALUE_LONG_NAME = "flow-fill-empty-bins-value";
     public static final String SYMMETRIC_INDELS_LONG_NAME = "flow-symmetric-indel-probs";
     public static final String REPORT_INS_OR_DEL_LONG_NAME = "flow-report-insertion-or-deletion";
     public static final String DISALLOW_LARGER_PROBS_LONG_NAME = "flow-disallow-probs-larger-than-call";
     public static final String LUMP_PROBS_LONG_NAME = "flow-lump-probs";
-    public static final String PROB_SF_LONG_NAME = "flow-probability-scaling-factor";
     public static final String RETAIN_MAX_N_PROBS_BASE_LONG_NAME = "flow-retain-max-n-probs-base-format";
-    public static final String FLOW_ORDER_CYCLE_LENGTH_LONG_NAME = "flow-order-cycle-length";
     public static final String FLOW_MATRIX_MODS_LONG_NAME = "flow-matrix-mods";
     public static final String FLOW_KEEP_BOUNDARY_FLOWS_LONG_NAME = "keep-boundary-flows";
 
@@ -32,13 +29,10 @@ public class FlowBasedArgumentCollection implements Serializable {
     private static final boolean DEFAULT_REMOVE_LONGER_INDELS = false;
     private static final boolean DEFAULT_REMOVE_ONE_TO_ZERO = false;
     private static final boolean DEFAULT_SYMMETRIC_INDELS = false;
-    private static final int DEFAULT_QUANTIZATION = 121;
     private static final boolean DEFAULT_ONLY_INS_OR_DEL = false;
     private static final boolean DEFAULT_DISALLOW_LARGER_PROBS = false;
     private static final boolean DEFAULT_LUMP_PROBS = false;
     private static final boolean DEFAULT_RETAIN_MAX_N_PROBS = false;
-    private static final int DEFAULT_PROB_SCALING_FACTOR = 10;
-    private static final int DEFAULT_FLOW_ORDER_CYCLE_LENGTH = 4;
     private static final boolean DEFAULT_FLOW_USE_T0_TAG = false;
 
     @Advanced
@@ -56,10 +50,6 @@ public class FlowBasedArgumentCollection implements Serializable {
     @Advanced
     @Argument(fullName = REMOVE_ONE_TO_ZERO_PROBS_LONG_NAME, doc = "Remove probabilities of basecall of zero from non-zero genome", optional = true)
     public boolean removeOneToZeroProbs = DEFAULT_REMOVE_ONE_TO_ZERO;
-
-    @Advanced
-    @Argument(fullName = NUMBER_OF_POSSIBLE_PROBS_LONG_NAME, doc = "Number of bins for probability quantization", optional = true)
-    public int probabilityQuantization = DEFAULT_QUANTIZATION;
 
     @Advanced
     @Argument(fullName = FILLING_VALUE_LONG_NAME, doc = "Value to fill the zeros of the matrix with", optional=true)
@@ -84,15 +74,6 @@ public class FlowBasedArgumentCollection implements Serializable {
     @Advanced
     @Argument(fullName = RETAIN_MAX_N_PROBS_BASE_LONG_NAME, doc = "Keep only hmer/2 probabilities (like in base format)", optional=true)
     public boolean retainMaxNProbs = DEFAULT_RETAIN_MAX_N_PROBS;
-
-    @Advanced
-    @Argument(fullName = PROB_SF_LONG_NAME, doc = "probability scaling factor for (phred=10) for probability quantization", optional=true)
-    public int probabilityScalingFactor = DEFAULT_PROB_SCALING_FACTOR;
-
-    @Advanced
-    @Hidden
-    @Argument(fullName = FLOW_ORDER_CYCLE_LENGTH_LONG_NAME, doc = "Length of flow order cycle", optional=true)
-    public int flowOrderCycleLength = DEFAULT_FLOW_ORDER_CYCLE_LENGTH;
 
     @Advanced
     @Argument(fullName=FLOW_MATRIX_MODS_LONG_NAME, doc="Modifications instructions to the read flow matrix. " +
